+++ client/CHANGELOG.md
@@ -1,13 +1,13 @@
 # Change Log
 
-All notable changes to the "WebRelease Template Language Support" extension will be documented in this file.
+All notable changes to the "WebRelease2 Template Language Support" extension will be documented in this file.
 
 ## [0.1.0] - 2024-12-16
 
 ### Added
 - Initial release
-- Syntax highlighting for WebRelease template language
+- Syntax highlighting for WebRelease2 template language
 - Syntax validation for expressions and tags
 - Code completion for tags, attributes, and functions
 - Hover information for tags and functions
-- Support for `.wr` and `.wrt` file extensions
+- Support for `.wr2` and `.wrt` file extensions